requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics==8.0.196
opencv-python>=4.7.0
numpy>=1.24.0
gradio>=3.50.0
pillow>=9.5.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
roboflow>=1.1.1
supervision>=0.19.0